refactor(terminalEngine): split processCommand into command modules (THI-59)
Extract ~1100 lines from the monolithic processCommand switch into four
dedicated modules under src/app/data/commands/:

- types.ts     — shared TerminalState/CommandOutput/FSNode types (no deps)
- git.ts       — full git simulation (handleGit)
- network.ts   — ping/curl/wget/ssh/scp/dig/nslookup (handleNetwork)
- env.ts       — export/env/printenv/source/crontab (handleEnv)
- windows.ts   — PowerShell aliases + macOS commands (handleWindows)

terminalEngine.ts: 2766 → 1680 lines (-39%)
processCommand switch: ~1300 → ~381 lines (-71%)
All types re-exported from terminalEngine for backward compat.
Windows handler uses dependency injection (WindowsCmdDeps) to avoid
circular imports with filesystem commands.

579/579 tests passing, 0 TS errors.
