+++ src/app/data/terminalEngine.ts
@@ -1,81 +1,13 @@
-// ─── Types ────────────────────────────────────────────────────────────────────
+// Types are defined in ./commands/types.ts and re-exported here for backward compatibility.
+import type { TerminalEnv, FileNode, DirectoryNode, FSNode, GitCommit, GitState, TerminalState, CommandOutput, OutputLine } from './commands/types';
+export type { TerminalEnv, FileNode, DirectoryNode, FSNode, GitCommit, GitState, TerminalState, CommandOutput, OutputLine };
 
-export type TerminalEnv = 'linux' | 'macos' | 'windows';
-
-export interface FileNode {
-  type: 'file';
-  content: string;
-  permissions: string;
-  owner: string;
-  group: string;
-  size: number;
-}
-
-export interface DirectoryNode {
-  type: 'directory';
-  children: Record<string, FSNode>;
-  permissions: string;
-  owner: string;
-  group: string;
-}
-
-export type FSNode = FileNode | DirectoryNode;
-
-// ─── Git Simulation Types ─────────────────────────────────────────────────────
-
-export interface GitCommit {
-  hash: string;
-  message: string;
-  author: string;
-  date: string;
-}
-
-/**
- * In-memory git state for the terminal simulator.
- * Tracks repo initialization, branches, staging area, history, and remotes.
- * Intentionally simplified — models the concepts taught in Modules 9 & 10,
- * not a full git implementation.
- */
-export interface GitState {
-  /** Whether `git init` has been run in the current directory */
-  initialized: boolean;
-  /** Currently checked-out branch name */
-  branch: string;
-  /** All local branches */
-  branches: string[];
-  /** Files staged for the next commit (relative paths as strings) */
-  stagedFiles: string[];
-  /** Commit history, newest first */
-  commits: GitCommit[];
-  /** Remote name → URL mapping */
-  remotes: Record<string, string>;
-}
-
-export interface TerminalState {
-  root: DirectoryNode;
-  cwd: string[];
-  commandHistory: string[];
-  user: string;
-  hostname: string;
-  /** Environment variables (export VAR=value / $env:VAR = "value"). */
-  envVars: Record<string, string>;
-  /**
-   * Git simulation state. Undefined until `git init` is run.
-   * Persists across commands within the same terminal session.
-   */
-  git?: GitState;
-}
-
-export interface CommandOutput {
-  lines: OutputLine[];
-  clear?: boolean;
-  newState: TerminalState;
-}
-
-export interface OutputLine {
-  text: string;
-  type: 'output' | 'error' | 'success' | 'info';
-}
+// ─── Command module handlers ──────────────────────────────────────────────────
+import { handleGit } from './commands/git';
+import { handleNetwork } from './commands/network';
+import { cmdEnv, handleEnv } from './commands/env';
+import { handleWindows } from './commands/windows';
+import type { WindowsCmdDeps } from './commands/windows';
 
 // ─── Initial Filesystem ───────────────────────────────────────────────────────
 
@@ -721,60 +653,9 @@ function cmdChmod(state: TerminalState, args: string[]): { lines: OutputLine[]; 
 
 // ─── Environment Variable Commands ───────────────────────────────────────────
 
-function cmdExport(args: string[], state: TerminalState): { lines: OutputLine[]; envVars: Record<string, string> } {
-  if (args.length === 0) {
-    // `export` with no args — list all vars
-    const lines = Object.entries(state.envVars).map(([k, v]) => ({
-      text: `declare -x ${k}="${v}"`,
-      type: 'output' as const,
-    }));
-    return { lines, envVars: state.envVars };
-  }
-  const newEnv = { ...state.envVars };
-  const lines: OutputLine[] = [];
-  for (const arg of args) {
-    const eqIdx = arg.indexOf('=');
-    if (eqIdx === -1) {
-      // `export VAR` without value — just marks for export (no-op here)
-      lines.push({ text: '', type: 'output' });
-    } else {
-      const name = arg.slice(0, eqIdx);
-      const value = arg.slice(eqIdx + 1).replace(/^["']|["']$/g, '');
-      if (!/^[A-Za-z_][A-Za-z0-9_]*$/.test(name)) {
-        lines.push({ text: `export: '${name}': not a valid identifier`, type: 'error' });
-      } else {
-        newEnv[name] = value;
-        lines.push({ text: '', type: 'output' });
-      }
-    }
-  }
-  return { lines: lines.filter((l) => l.text !== ''), envVars: newEnv };
-}
+// cmdExport, cmdEnv, cmdPrintenv, cmdSource → moved to ./commands/env.ts
 
-function cmdEnv(state: TerminalState): OutputLine[] {
-  return Object.entries(state.envVars).map(([k, v]) => ({
-    text: `${k}=${v}`,
-    type: 'output' as const,
-  }));
-}
-
-function cmdPrintenv(args: string[], state: TerminalState): OutputLine[] {
-  if (args.length === 0) return cmdEnv(state);
-  return args.map((name) => {
-    if (Object.prototype.hasOwnProperty.call(state.envVars, name)) {
-      return { text: state.envVars[name], type: 'output' as const };
-    }
-    return { text: '', type: 'error' as const };
-  }).filter((l) => l.text !== '');
-}
-
-function cmdSource(args: string[]): OutputLine[] {
-  const file = args[0] ?? '(unknown)';
-  const display = file.replace('~/', '$HOME/');
-  return [{ text: `${display}: loaded`, type: 'success' }];
-}
-
-// ─── New Module 5/6 Commands ──────────────────────────────────────────────────
+// ─── Module 5/6 Commands ─────────────────────────────────────────────────────
 
 function cmdChown(args: string[]): OutputLine[] {
   if (args.length < 2) {
@@ -850,23 +731,7 @@ function cmdGetJob(): OutputLine[] {
   ];
 }
 
-function cmdCrontab(args: string[]): OutputLine[] {
-  if (args.includes('-l')) {
-    return [
-      { text: '# Crontab — tâches planifiées', type: 'output' },
-      { text: '# Format: minute heure jour mois jour_semaine commande', type: 'output' },
-      { text: '0 9 * * 1-5 /home/user/projets/script.sh   # Tous les jours de semaine à 9h', type: 'output' },
-      { text: '*/30 * * * * /usr/local/bin/backup.sh       # Toutes les 30 minutes', type: 'output' },
-    ];
-  }
-  if (args.includes('-e')) {
-    return [{ text: 'crontab: ouverture de l\'éditeur (nano). En mode simulé, utilisez "crontab -l" pour voir les entrées.', type: 'info' }];
-  }
-  if (args.includes('-r')) {
-    return [{ text: 'crontab supprimé.', type: 'success' }];
-  }
-  return [{ text: 'Usage: crontab [-l] [-e] [-r]', type: 'error' }];
-}
+// cmdCrontab → moved to ./commands/env.ts
 
 function cmdEchoRedirect(
   state: TerminalState,
@@ -1499,6 +1364,12 @@ export function processCommand(state: TerminalState, input: string, env: Termina
   const cmd = parts[0]?.toLowerCase();
   const args = parts.slice(1);
 
+  // Dependencies for Windows/macOS alias handler
+  const winDeps: WindowsCmdDeps = {
+    cmdPwd, cmdCd, cmdLs, cmdCat, cmdMkdir, cmdTouch,
+    cmdCp, cmdMv, cmdRm, cmdEcho, cmdGrep, cmdEnv,
+  };
+
   switch (cmd) {
     case 'pwd':
       return { lines: cmdPwd(newState, env), newState };
@@ -1530,24 +1401,14 @@ export function processCommand(state: TerminalState, input: string, env: Termina
     case 'echo':
       return { lines: cmdEcho(args, newState.envVars), newState };
 
-    case 'export': {
-      if (env === 'windows') return { lines: [{ text: 'Utilisez $env:VAR = "value" en PowerShell pour définir une variable.', type: 'info' }], newState };
-      const { lines: exportLines, envVars: newEnv } = cmdExport(args, newState);
-      return { lines: exportLines, newState: { ...newState, envVars: newEnv } };
-    }
-
+    // ── Environment & scripts → commands/env.ts ───────────────────────────────
+    case 'export':
     case 'env':
-      return { lines: cmdEnv(newState), newState };
-
     case 'printenv':
-      return { lines: cmdPrintenv(args, newState), newState };
-
     case 'source':
     case '.':
-      return { lines: cmdSource(args), newState };
-
     case 'crontab':
-      return { lines: cmdCrontab(args), newState };
+      return handleEnv(cmd, args, newState, env);
 
     case 'chown': {
       if (env === 'windows') return { lines: [{ text: 'chown n\'est pas disponible sur Windows. Utilisez icacls ou takeown.', type: 'info' }], newState };
@@ -1787,980 +1648,33 @@ export function processCommand(state: TerminalState, input: string, env: Termina
         newState,
       };
 
-    // ── Windows PowerShell aliases ─────────────────────────────────────────────
 
-    // pwd equivalents
-    case 'get-location':
-    case 'gl':
-      return { lines: cmdPwd(newState, env), newState };
-
-    // cd equivalents
-    case 'set-location':
-    case 'sl': {
-      const { lines: slLines, newCwd: slCwd } = cmdCd(newState, args);
-      if (slCwd) newState = { ...newState, cwd: slCwd };
-      return { lines: slLines, newState };
-    }
-
-    // ls equivalents (get-childitem also handles Env: provider)
-    case 'get-childitem':
-    case 'gci':
-    case 'dir':
-      if (args[0]?.toLowerCase() === 'env:') {
-        return { lines: cmdEnv(newState), newState };
-      }
-      return { lines: cmdLs(newState, args), newState };
-
-    // cat equivalents
-    case 'get-content':
-    case 'gc':
-      return { lines: cmdCat(newState, args), newState };
-
-    // New-Item: creates file or directory
-    // Usage: New-Item name | New-Item -ItemType Directory -Name name | New-Item -ItemType File -Name name
-    case 'new-item':
-    case 'ni': {
-      const isDir = args.some((a) => a.toLowerCase() === 'directory');
-      // -Name flag or first non-flag arg
-      const nameIdx = args.findIndex((a) => a.toLowerCase() === '-name');
-      const pathIdx = args.findIndex((a) => a.toLowerCase() === '-path');
-      const name =
-        nameIdx >= 0 ? args[nameIdx + 1] :
-        pathIdx >= 0 ? args[pathIdx + 1] :
-        args.find((a) => !a.startsWith('-'));
-      if (!name) return { lines: [{ text: 'New-Item: -Name ou chemin requis', type: 'error' }], newState };
-      if (isDir) {
-        const { lines: niLines, newRoot: niRoot } = cmdMkdir(newState, [name]);
-        if (niRoot) newState = { ...newState, root: niRoot };
-        return { lines: niLines, newState };
-      } else {
-        const { lines: niLines, newRoot: niRoot } = cmdTouch(newState, [name]);
-        if (niRoot) newState = { ...newState, root: niRoot };
-        return { lines: niLines, newState };
-      }
-    }
-
-    // cp equivalents
-    case 'copy-item':
-    case 'cpi':
-    case 'copy': {
-      const cpArgs = args.filter((a) => !a.startsWith('-'));
-      const { lines: cpLines, newRoot: cpRoot } = cmdCp(newState, cpArgs);
-      if (cpRoot) newState = { ...newState, root: cpRoot };
-      return { lines: cpLines, newState };
-    }
-
-    // mv equivalents
-    case 'move-item':
-    case 'mi':
-    case 'move': {
-      const mvArgs = args.filter((a) => !a.startsWith('-'));
-      const { lines: mvLines, newRoot: mvRoot } = cmdMv(newState, mvArgs);
-      if (mvRoot) newState = { ...newState, root: mvRoot };
-      return { lines: mvLines, newState };
-    }
-
-    // rm equivalents
-    case 'remove-item':
-    case 'ri':
-    case 'del':
-    case 'erase': {
-      const rmArgs = args.filter((a) => !a.startsWith('-'));
-      const { lines: rmLines, newRoot: rmRoot } = cmdRm(newState, rmArgs);
-      if (rmRoot) newState = { ...newState, root: rmRoot };
-      return { lines: rmLines, newState };
-    }
-
-    // echo equivalents
-    case 'write-host':
-    case 'write-output':
-      return { lines: cmdEcho(args.filter((a) => !a.startsWith('-')), newState.envVars), newState };
-
-    // PowerShell env var listing — handled inline in get-childitem below
-
-    // ps equivalents
-    case 'get-process':
-    case 'gps': {
-      const psLines = [
-        { text: 'Handles  NPM(K)  PM(K)  WS(K) VM(M)   CPU(s)   Id ProcessName', type: 'output' as const },
-        { text: '-------  ------  -----  ----- -----   ------   -- -----------', type: 'output' as const },
-        { text: `    256      14   4560   8192   120    0.047 1234 WindowsTerminal`, type: 'output' as const },
-        { text: `     64       8   2048   4096    80    0.016 2048 node`, type: 'output' as const },
-        { text: `     32       4   1024   2048    40    0.003 5678 pwsh`, type: 'output' as const },
-      ];
-      return { lines: psLines, newState };
-    }
-
-    // kill equivalents
-    case 'stop-process':
-    case 'spps':
-    case 'taskkill': {
-      if (!args.length) return { lines: [{ text: 'Stop-Process: -Id ou -Name requis', type: 'error' }], newState };
-      const target = args[args.length - 1];
-      return { lines: [{ text: `Processus '${target}' arrêté.`, type: 'success' }], newState };
-    }
-
-    // grep equivalents
-    // Select-String -Pattern "pattern" -Path file  OR  Select-String "pattern" file
-    case 'select-string':
-    case 'sls': {
-      const patternIdx = args.findIndex((a) => a.toLowerCase() === '-pattern');
-      const pathIdx = args.findIndex((a) => a.toLowerCase() === '-path');
-      let pattern: string;
-      let filePath: string;
-      if (patternIdx >= 0) {
-        pattern = args[patternIdx + 1] ?? '';
-        filePath = pathIdx >= 0 ? args[pathIdx + 1] : (args.find((a, i) => !a.startsWith('-') && i !== patternIdx + 1) ?? '');
-      } else {
-        const nonFlags = args.filter((a) => !a.startsWith('-'));
-        pattern = nonFlags[0] ?? '';
-        filePath = nonFlags[1] ?? '';
-      }
-      return { lines: cmdGrep(newState, [pattern, filePath]), newState };
-    }
-
-    // clear equivalents
-    case 'clear-host':
-    case 'cls':
-      return { lines: [], clear: true, newState };
-
-    // mkdir alias (works on all OS)
-    case 'md':
-      return (() => {
-        const { lines: mdLines, newRoot: mdRoot } = cmdMkdir(newState, args);
-        if (mdRoot) newState = { ...newState, root: mdRoot };
-        return { lines: mdLines, newState };
-      })();
-
-    // ── macOS-specific ─────────────────────────────────────────────────────────
-
-    // open: simulates opening a file/app
-    case 'open': {
-      if (!args.length) return { lines: [{ text: 'open: missing argument', type: 'error' }], newState };
-      const target = args[args.length - 1];
-      return { lines: [{ text: `Ouverture de '${target}'…`, type: 'success' }], newState };
-    }
-
-    // pbcopy: copy stdin to clipboard (simulated)
-    case 'pbcopy':
-      return { lines: [{ text: '[contenu copié dans le presse-papiers]', type: 'success' }], newState };
-
-    // pbpaste: paste clipboard content (simulated)
-    case 'pbpaste':
-      return { lines: [{ text: '[contenu du presse-papiers]', type: 'output' }], newState };
-
-    // brew: Homebrew package manager (simulated)
-    case 'brew': {
-      const sub = args[0]?.toLowerCase();
-      if (sub === 'install' && args[1]) {
-        return { lines: [
-          { text: `==> Fetching ${args[1]}...`, type: 'info' },
-          { text: `==> Installing ${args[1]}...`, type: 'info' },
-          { text: `✓  ${args[1]} installed successfully`, type: 'success' },
-        ], newState };
-      }
-      if (sub === 'update') {
-        return { lines: [{ text: '==> Updated Homebrew. Nothing to upgrade.', type: 'success' }], newState };
-      }
-      if (sub === 'list') {
-        return { lines: [{ text: 'git  node  python  wget  curl', type: 'output' }], newState };
-      }
-      return { lines: [{ text: 'Homebrew — Usage: brew install|update|list', type: 'output' }], newState };
-    }
-
-    // winget: Windows package manager (simulated)
-    case 'winget': {
-      const sub = args[0]?.toLowerCase();
-      if (sub === 'install' && args[1]) {
-        return { lines: [
-          { text: `Found ${args[1]}`, type: 'info' },
-          { text: `Downloading ${args[1]}...`, type: 'info' },
-          { text: `Successfully installed ${args[1]}`, type: 'success' },
-        ], newState };
-      }
-      if (sub === 'list') {
-        return { lines: [{ text: 'Name          Id                  Version', type: 'output' },
-          { text: 'Git           Git.Git             2.44.0', type: 'output' },
-          { text: 'Node.js       OpenJS.NodeJS       20.11.0', type: 'output' },
-        ], newState };
-      }
-      return { lines: [{ text: 'winget — Usage: winget install|list', type: 'output' }], newState };
-    }
-
-    // ── Réseau & SSH (Module 8) ───────────────────────────────────────────────
-
-    case 'ping': {
-      const host = args.find((a) => !a.startsWith('-') && isNaN(Number(a))) ?? '';
-      if (!host) return { lines: [{ text: 'Usage: ping <hostname>', type: 'error' }], newState };
-      return {
-        lines: [
-          { text: `PING ${host}: 56 data bytes`, type: 'output' },
-          { text: `64 bytes from ${host}: icmp_seq=0 ttl=54 time=12.3 ms`, type: 'output' },
-          { text: `64 bytes from ${host}: icmp_seq=1 ttl=54 time=11.8 ms`, type: 'output' },
-          { text: `64 bytes from ${host}: icmp_seq=2 ttl=54 time=12.1 ms`, type: 'output' },
-          { text: `--- ${host} ping statistics ---`, type: 'output' },
-          { text: '3 packets transmitted, 3 received, 0% packet loss', type: 'success' },
-        ],
-        newState,
-      };
-    }
-
-    case 'curl': {
-      const url = args.find((a) => !a.startsWith('-')) ?? '';
-      if (!url) return { lines: [{ text: 'Usage: curl [options] <url>', type: 'error' }], newState };
-      const urlHost = url.replace(/^https?:\/\//, '').split('/')[0] ?? 'server';
-      if (args[0] === '-I' || args[0] === '--head') {
-        return {
-          lines: [
-            { text: 'HTTP/2 200', type: 'success' },
-            { text: 'content-type: application/json; charset=utf-8', type: 'output' },
-            { text: `server: ${urlHost}`, type: 'output' },
-            { text: 'x-content-type-options: nosniff', type: 'output' },
-          ],
-          newState,
-        };
-      }
-      return {
-        lines: [
-          { text: `{"url":"${url}","status":"ok"}`, type: 'output' },
-        ],
-        newState,
-      };
-    }
-
-    case 'wget': {
-      const url = args.find((a) => !a.startsWith('-')) ?? '';
-      if (!url) return { lines: [{ text: 'Usage: wget <url>', type: 'error' }], newState };
-      const filename = url.split('/').pop() || 'index.html';
-      return {
-        lines: [
-          { text: `Connecting to ${url.split('/')[2] ?? 'host'}... connected.`, type: 'output' },
-          { text: 'HTTP request sent, awaiting response... 200 OK', type: 'success' },
-          { text: `Saving to: '${filename}'`, type: 'output' },
-          { text: `'${filename}' saved [1024]`, type: 'success' },
-        ],
-        newState,
-      };
-    }
-
+    // ── Network & SSH (Module 8) → commands/network.ts ──────────────────────
+    case 'ping':
+    case 'curl':
+    case 'wget':
     case 'invoke-webrequest':
-    case 'iwr': {
-      const url = args.find((a) => !a.startsWith('-')) ?? '';
-      if (!url) return { lines: [{ text: 'Usage: Invoke-WebRequest -Uri <url>', type: 'error' }], newState };
-      const outFile = (() => { const i = args.indexOf('-OutFile'); return i >= 0 ? args[i + 1] : null; })();
-      if (outFile) {
-        return {
-          lines: [
-            { text: `Downloading ${url}...`, type: 'output' },
-            { text: `Content saved to '${outFile}'`, type: 'success' },
-          ],
-          newState,
-        };
-      }
-      return {
-        lines: [
-          { text: 'StatusCode        : 200', type: 'output' },
-          { text: 'StatusDescription : OK', type: 'output' },
-          { text: `Content           : {"url":"${url}","status":"ok"}`, type: 'output' },
-        ],
-        newState,
-      };
-    }
+    case 'iwr':
+    case 'nslookup':
+    case 'dig':
+    case 'resolve-dnsname':
+    case 'ssh':
+    case 'ssh-keygen':
+    case 'scp':
+      return handleNetwork(cmd, args, newState);
 
-    case 'nslookup': {
-      const host = args[0] ?? '';
-      if (!host) return { lines: [{ text: 'Usage: nslookup <hostname>', type: 'error' }], newState };
-      return {
-        lines: [
-          { text: 'Server:  8.8.8.8', type: 'output' },
-          { text: 'Address: 8.8.8.8#53', type: 'output' },
-          { text: 'Non-authoritative answer:', type: 'output' },
-          { text: `Name: ${host}`, type: 'output' },
-          { text: 'Address: 142.250.74.46', type: 'output' },
-        ],
-        newState,
-      };
-    }
+    // ── Git (Modules 9 & 10) → commands/git.ts ───────────────────────────────
+    case 'git':
+      return handleGit(newState, args, env);
 
-    case 'dig': {
-      const host = args.find((a) => !a.startsWith('+') && !a.startsWith('@')) ?? '';
-      if (!host) return { lines: [{ text: 'Usage: dig <hostname>', type: 'error' }], newState };
-      return {
-        lines: [
-          { text: `; <<>> DiG 9.18.1 <<>> ${host}`, type: 'output' },
-          { text: ';; ANSWER SECTION:', type: 'output' },
-          { text: `${host}. 299 IN A 142.250.74.46`, type: 'output' },
-          { text: ';; Query time: 12 msec', type: 'output' },
-          { text: ';; SERVER: 8.8.8.8#53', type: 'output' },
-        ],
-        newState,
-      };
-    }
-
-    case 'resolve-dnsname': {
-      const host = args.find((a) => !a.startsWith('-')) ?? '';
-      if (!host) return { lines: [{ text: 'Usage: Resolve-DnsName <hostname>', type: 'error' }], newState };
-      return {
-        lines: [
-          { text: 'Name                           Type TTL  Section IPAddress', type: 'output' },
-          { text: '----                           ---- ---  ------- ---------', type: 'output' },
-          { text: `${host.padEnd(31)}A    299  Answer  142.250.74.46`, type: 'output' },
-        ],
-        newState,
-      };
-    }
-
-    case 'ssh': {
-      const target = args.find((a) => !a.startsWith('-')) ?? '';
-      if (!target) return { lines: [{ text: 'Usage: ssh user@hostname', type: 'error' }], newState };
-      return {
-        lines: [
-          { text: `ssh: connexion simulée vers ${target}`, type: 'info' },
-          { text: '(Dans un vrai terminal, vous seriez connecté à l\'hôte distant)', type: 'info' },
-        ],
-        newState,
-      };
-    }
-
-    case 'ssh-keygen': {
-      const tIdx = args.indexOf('-t');
-      const keyType = tIdx >= 0 ? (args[tIdx + 1] ?? 'rsa') : 'rsa';
-      return {
-        lines: [
-          { text: `Generating public/private ${keyType} key pair.`, type: 'output' },
-          { text: `Enter file in which to save the key (/home/user/.ssh/id_${keyType}): (simulé)`, type: 'output' },
-          { text: `Your identification has been saved in /home/user/.ssh/id_${keyType}`, type: 'success' },
-          { text: `Your public key has been saved in /home/user/.ssh/id_${keyType}.pub`, type: 'success' },
-          { text: `+--[${keyType.toUpperCase()}]--+`, type: 'output' },
-          { text: '|     .o+.        |', type: 'output' },
-          { text: '+----[SHA256]-----+', type: 'output' },
-        ],
-        newState,
-      };
-    }
-
-    case 'scp': {
-      if (args.length < 2) return { lines: [{ text: 'Usage: scp <source> user@host:<dest>', type: 'error' }], newState };
-      const src = args.find((a) => !a.startsWith('-') && !a.includes('@')) ?? args[0];
-      return {
-        lines: [
-          { text: `${src}      100%  1024   512.0KB/s   00:00`, type: 'success' },
-        ],
-        newState,
-      };
-    }
-
-    // ── Git (Modules 9 & 10) ──────────────────────────────────────────────────
-    // Full git simulation: init, add, commit, status, log, diff, branch,
-    // checkout, merge, remote, push, pull, fetch, clone.
-    // State is persisted in TerminalState.git across commands.
-
-    case 'git': {
-      const sub = args[0]?.toLowerCase() ?? '';
-
-      /** Require an initialised repo for sub-commands that need one. */
-      const requireRepo = (): OutputLine | null => {
-        if (!newState.git?.initialized) {
-          return { text: "fatal: not a git repository (or any parent up to mount point /)\nHint: Run 'git init' to create a new repository.", type: 'error' };
-        }
-        return null;
-      };
-
-      /** Generate a short 7-char hex hash. */
-      const makeHash = (): string =>
-        Array.from({ length: 7 }, () => '0123456789abcdef'[Math.floor(Math.random() * 16)]).join('');
-
-      switch (sub) {
-        // ── git init ──────────────────────────────────────────────────────────
-        case 'init': {
-          if (newState.git?.initialized) {
-            return {
-              lines: [{ text: `Reinitialized existing Git repository in ${displayPath(newState.cwd)}/.git/`, type: 'info' }],
-              newState,
-            };
-          }
-          newState = {
-            ...newState,
-            git: {
-              initialized: true,
-              branch: 'main',
-              branches: ['main'],
-              stagedFiles: [],
-              commits: [],
-              remotes: {},
-            },
-          };
-          return {
-            lines: [
-              { text: `Initialized empty Git repository in ${displayPath(newState.cwd)}/.git/`, type: 'success' },
-              { text: "Hint: Use 'git add <file>' to stage files, 'git commit -m' to record changes.", type: 'info' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git status ────────────────────────────────────────────────────────
-        case 'status': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const lines: OutputLine[] = [
-            { text: `On branch ${g.branch}`, type: 'output' },
-          ];
-          if (g.commits.length === 0) {
-            lines.push({ text: 'No commits yet', type: 'output' });
-          }
-          if (g.stagedFiles.length > 0) {
-            lines.push({ text: '', type: 'output' });
-            lines.push({ text: 'Changes to be committed:', type: 'success' });
-            lines.push({ text: '  (use "git restore --staged <file>" to unstage)', type: 'info' });
-            g.stagedFiles.forEach((f) => lines.push({ text: `\tnew file:   ${f}`, type: 'success' }));
-          }
-          if (g.stagedFiles.length === 0 && g.commits.length === 0) {
-            lines.push({ text: '', type: 'output' });
-            lines.push({ text: 'Nothing to commit.', type: 'output' });
-            lines.push({ text: "Hint: Stage files with 'git add <file>' or 'git add .'", type: 'info' });
-          }
-          if (g.stagedFiles.length === 0 && g.commits.length > 0) {
-            lines.push({ text: '', type: 'output' });
-            lines.push({ text: 'nothing to commit, working tree clean', type: 'output' });
-          }
-          return { lines, newState };
-        }
-
-        // ── git add ───────────────────────────────────────────────────────────
-        case 'add': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const target = args[1] ?? '';
-          if (!target) {
-            return { lines: [{ text: 'Nothing specified, nothing added.\nHint: git add <file> or git add .', type: 'error' }], newState };
-          }
-          const g = newState.git!;
-          let filesToStage: string[];
-          if (target === '.' || target === '-A' || target === '--all') {
-            // Stage all visible files in cwd (simulated)
-            const cwdNode = getNode(newState.root, newState.cwd);
-            if (cwdNode?.type === 'directory') {
-              filesToStage = Object.keys(cwdNode.children)
-                .filter((n) => !n.startsWith('.') && !g.stagedFiles.includes(n));
-            } else {
-              filesToStage = [];
-            }
-          } else {
-            filesToStage = [target].filter((f) => !g.stagedFiles.includes(f));
-          }
-          if (filesToStage.length === 0) {
-            return { lines: [{ text: `'${target}' already staged or no files to add.`, type: 'info' }], newState };
-          }
-          newState = { ...newState, git: { ...g, stagedFiles: [...g.stagedFiles, ...filesToStage] } };
-          return {
-            lines: filesToStage.map((f) => ({ text: `staged: ${f}`, type: 'success' })),
-            newState,
-          };
-        }
-
-        // ── git restore ───────────────────────────────────────────────────────
-        case 'restore': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const staged = args.includes('--staged');
-          const file = args.find((a) => !a.startsWith('-')) ?? '';
-          if (!file) return { lines: [{ text: 'Usage: git restore [--staged] <file>', type: 'error' }], newState };
-          if (staged) {
-            const g = newState.git!;
-            newState = { ...newState, git: { ...g, stagedFiles: g.stagedFiles.filter((f) => f !== file) } };
-            return { lines: [{ text: `unstaged: ${file}`, type: 'info' }], newState };
-          }
-          return { lines: [{ text: `Restored '${file}' (simulated)`, type: 'info' }], newState };
-        }
-
-        // ── git commit ────────────────────────────────────────────────────────
-        case 'commit': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const mIdx = args.indexOf('-m');
-          const message = mIdx >= 0 ? (args[mIdx + 1] ?? '') : '';
-          if (!message) {
-            return { lines: [{ text: 'Abort: empty commit message.\nUsage: git commit -m "your message"', type: 'error' }], newState };
-          }
-          if (g.stagedFiles.length === 0) {
-            return { lines: [{ text: 'nothing to commit, working tree clean', type: 'info' }], newState };
-          }
-          const hash = makeHash();
-          const commit: GitCommit = {
-            hash,
-            message,
-            author: newState.user,
-            date: new Date().toISOString().slice(0, 10),
-          };
-          const fileCount = g.stagedFiles.length;
-          newState = {
-            ...newState,
-            git: { ...g, commits: [commit, ...g.commits], stagedFiles: [] },
-          };
-          return {
-            lines: [
-              { text: `[${g.branch} ${hash}] ${message}`, type: 'success' },
-              { text: ` ${fileCount} file${fileCount !== 1 ? 's' : ''} changed`, type: 'output' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git log ───────────────────────────────────────────────────────────
-        case 'log': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          if (g.commits.length === 0) {
-            return { lines: [{ text: 'fatal: your current branch has no commits yet.', type: 'error' }], newState };
-          }
-          const oneline = args.includes('--oneline');
-          const lines: OutputLine[] = [];
-          g.commits.forEach((c) => {
-            if (oneline) {
-              lines.push({ text: `${c.hash} ${c.message}`, type: 'output' });
-            } else {
-              lines.push({ text: `commit ${c.hash}`, type: 'success' });
-              lines.push({ text: `Author: ${c.author} <${c.author}@terminal-lab.local>`, type: 'output' });
-              lines.push({ text: `Date:   ${c.date}`, type: 'output' });
-              lines.push({ text: '', type: 'output' });
-              lines.push({ text: `    ${c.message}`, type: 'output' });
-              lines.push({ text: '', type: 'output' });
-            }
-          });
-          return { lines, newState };
-        }
-
-        // ── git diff ──────────────────────────────────────────────────────────
-        case 'diff': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          if (g.stagedFiles.length === 0 && g.commits.length > 0) {
-            return { lines: [{ text: '(nothing to diff — working tree clean)', type: 'info' }], newState };
-          }
-          return {
-            lines: [
-              { text: 'diff --git a/fichier.txt b/fichier.txt', type: 'output' },
-              { text: '--- a/fichier.txt', type: 'output' },
-              { text: '+++ b/fichier.txt', type: 'output' },
-              { text: '@@ -1,3 +1,4 @@', type: 'info' },
-              { text: ' Ligne existante', type: 'output' },
-              { text: '+Nouvelle ligne ajoutée', type: 'success' },
-              { text: '-Ligne supprimée', type: 'error' },
-              { text: ' Autre ligne', type: 'output' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git branch ────────────────────────────────────────────────────────
-        case 'branch': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const deleteFlag = args.includes('-d') || args.includes('-D');
-          const branchName = args.find((a) => !a.startsWith('-') && a !== sub) ?? '';
-
-          // List branches
-          if (!branchName && !deleteFlag) {
-            return {
-              lines: g.branches.map((b) => ({
-                text: b === g.branch ? `* ${b}` : `  ${b}`,
-                type: b === g.branch ? 'success' : 'output',
-              })),
-              newState,
-            };
-          }
-
-          // Delete branch
-          if (deleteFlag && branchName) {
-            if (branchName === g.branch) {
-              return { lines: [{ text: `error: Cannot delete branch '${branchName}' checked out at current worktree.`, type: 'error' }], newState };
-            }
-            if (!g.branches.includes(branchName)) {
-              return { lines: [{ text: `error: branch '${branchName}' not found.`, type: 'error' }], newState };
-            }
-            newState = { ...newState, git: { ...g, branches: g.branches.filter((b) => b !== branchName) } };
-            return { lines: [{ text: `Deleted branch ${branchName}.`, type: 'success' }], newState };
-          }
-
-          // Create branch
-          if (branchName) {
-            if (g.branches.includes(branchName)) {
-              return { lines: [{ text: `fatal: A branch named '${branchName}' already exists.`, type: 'error' }], newState };
-            }
-            newState = { ...newState, git: { ...g, branches: [...g.branches, branchName] } };
-            return { lines: [{ text: `Branch '${branchName}' created.`, type: 'success' }], newState };
-          }
-
-          return { lines: [{ text: 'Usage: git branch [<branch-name>] [-d <branch>]', type: 'error' }], newState };
-        }
-
-        // ── git checkout ──────────────────────────────────────────────────────
-        case 'checkout': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const createFlag = args.includes('-b') || args.includes('-B');
-          const branchName = args.find((a) => !a.startsWith('-') && a !== sub) ?? '';
-
-          if (!branchName) {
-            return { lines: [{ text: 'Usage: git checkout [-b] <branch>', type: 'error' }], newState };
-          }
-
-          if (createFlag) {
-            if (g.branches.includes(branchName)) {
-              return { lines: [{ text: `fatal: A branch named '${branchName}' already exists.`, type: 'error' }], newState };
-            }
-            newState = { ...newState, git: { ...g, branch: branchName, branches: [...g.branches, branchName] } };
-            return { lines: [{ text: `Switched to a new branch '${branchName}'`, type: 'success' }], newState };
-          }
-
-          if (!g.branches.includes(branchName)) {
-            return { lines: [{ text: `error: pathspec '${branchName}' did not match any branch known to git.`, type: 'error' }], newState };
-          }
-          if (branchName === g.branch) {
-            return { lines: [{ text: `Already on '${branchName}'`, type: 'info' }], newState };
-          }
-          newState = { ...newState, git: { ...g, branch: branchName } };
-          return { lines: [{ text: `Switched to branch '${branchName}'`, type: 'success' }], newState };
-        }
-
-        // ── git switch (modern alias for checkout) ────────────────────────────
-        case 'switch': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const createFlag = args.includes('-c') || args.includes('-C');
-          const branchName = args.find((a) => !a.startsWith('-') && a !== sub) ?? '';
-
-          if (!branchName) return { lines: [{ text: 'Usage: git switch [-c] <branch>', type: 'error' }], newState };
-
-          if (createFlag) {
-            if (g.branches.includes(branchName)) {
-              return { lines: [{ text: `fatal: A branch named '${branchName}' already exists.`, type: 'error' }], newState };
-            }
-            newState = { ...newState, git: { ...g, branch: branchName, branches: [...g.branches, branchName] } };
-            return { lines: [{ text: `Switched to a new branch '${branchName}'`, type: 'success' }], newState };
-          }
-          if (!g.branches.includes(branchName)) {
-            return { lines: [{ text: `fatal: invalid reference: ${branchName}`, type: 'error' }], newState };
-          }
-          newState = { ...newState, git: { ...g, branch: branchName } };
-          return { lines: [{ text: `Switched to branch '${branchName}'`, type: 'success' }], newState };
-        }
-
-        // ── git merge ─────────────────────────────────────────────────────────
-        case 'merge': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const branchName = args.find((a) => !a.startsWith('-') && a !== sub) ?? '';
-          if (!branchName) return { lines: [{ text: 'Usage: git merge <branch>', type: 'error' }], newState };
-          if (!g.branches.includes(branchName)) {
-            return { lines: [{ text: `merge: ${branchName} - not something we can merge`, type: 'error' }], newState };
-          }
-          if (branchName === g.branch) {
-            return { lines: [{ text: 'Already up to date.', type: 'info' }], newState };
-          }
-          const hash = makeHash();
-          const mergeCommit: GitCommit = {
-            hash,
-            message: `Merge branch '${branchName}' into ${g.branch}`,
-            author: newState.user,
-            date: new Date().toISOString().slice(0, 10),
-          };
-          newState = { ...newState, git: { ...g, commits: [mergeCommit, ...g.commits] } };
-          return {
-            lines: [
-              { text: `Merge made by the 'ort' strategy.`, type: 'success' },
-              { text: `[${g.branch} ${hash}] Merge branch '${branchName}'`, type: 'output' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git remote ────────────────────────────────────────────────────────
-        case 'remote': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const remoteSub = args[1]?.toLowerCase() ?? '';
-
-          if (!remoteSub || remoteSub === '-v' || remoteSub === '--verbose') {
-            if (Object.keys(g.remotes).length === 0) {
-              return { lines: [{ text: '(no remotes configured)', type: 'info' }], newState };
-            }
-            const lines: OutputLine[] = [];
-            Object.entries(g.remotes).forEach(([name, url]) => {
-              lines.push({ text: `${name}\t${url} (fetch)`, type: 'output' });
-              lines.push({ text: `${name}\t${url} (push)`, type: 'output' });
-            });
-            return { lines, newState };
-          }
-
-          if (remoteSub === 'add') {
-            const name = args[2] ?? '';
-            const url = args[3] ?? '';
-            if (!name || !url) return { lines: [{ text: 'Usage: git remote add <name> <url>', type: 'error' }], newState };
-            if (g.remotes[name]) return { lines: [{ text: `error: remote '${name}' already exists.`, type: 'error' }], newState };
-            newState = { ...newState, git: { ...g, remotes: { ...g.remotes, [name]: url } } };
-            return { lines: [{ text: `Remote '${name}' added (${url})`, type: 'success' }], newState };
-          }
-
-          if (remoteSub === 'remove' || remoteSub === 'rm') {
-            const name = args[2] ?? '';
-            if (!name) return { lines: [{ text: 'Usage: git remote remove <name>', type: 'error' }], newState };
-            if (!g.remotes[name]) return { lines: [{ text: `error: No such remote '${name}'`, type: 'error' }], newState };
-            // eslint-disable-next-line @typescript-eslint/no-unused-vars
-            const { [name]: _r, ...rest } = g.remotes;
-            newState = { ...newState, git: { ...g, remotes: rest } };
-            return { lines: [{ text: `Remote '${name}' removed.`, type: 'success' }], newState };
-          }
-
-          return { lines: [{ text: 'Usage: git remote add|remove|[-v]', type: 'error' }], newState };
-        }
-
-        // ── git push ──────────────────────────────────────────────────────────
-        case 'push': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const hasRemote = Object.keys(g.remotes).length > 0;
-          const remoteName = args.find((a) => !a.startsWith('-') && a !== sub) ?? 'origin';
-          if (!hasRemote) {
-            return { lines: [{ text: `fatal: '${remoteName}' does not appear to be a git repository.\nHint: git remote add origin <url>`, type: 'error' }], newState };
-          }
-          if (g.commits.length === 0) {
-            return { lines: [{ text: 'Everything up-to-date (no commits to push).', type: 'info' }], newState };
-          }
-          const upstreamFlag = args.includes('-u') || args.includes('--set-upstream');
-          const lines: OutputLine[] = [
-            { text: `Enumerating objects: ${g.commits.length}, done.`, type: 'output' },
-            { text: `Counting objects: 100% (${g.commits.length}/${g.commits.length}), done.`, type: 'output' },
-            { text: `Writing objects: 100% (${g.commits.length}/${g.commits.length}), done.`, type: 'output' },
-            { text: `To ${Object.values(g.remotes)[0]}`, type: 'output' },
-            { text: `   ${g.commits[g.commits.length - 1]?.hash ?? '0000000'}..${g.commits[0]?.hash ?? '0000000'}  ${g.branch} -> ${g.branch}`, type: 'success' },
-          ];
-          if (upstreamFlag) {
-            lines.push({ text: `Branch '${g.branch}' set up to track remote branch '${g.branch}' from '${remoteName}'.`, type: 'success' });
-          }
-          return { lines, newState };
-        }
-
-        // ── git pull ──────────────────────────────────────────────────────────
-        case 'pull': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          if (Object.keys(g.remotes).length === 0) {
-            return { lines: [{ text: "There is no tracking information for the current branch.\nHint: git remote add origin <url>", type: 'error' }], newState };
-          }
-          return {
-            lines: [
-              { text: 'remote: Enumerating objects: 3, done.', type: 'output' },
-              { text: 'remote: Counting objects: 100% (3/3), done.', type: 'output' },
-              { text: 'Updating... Fast-forward', type: 'output' },
-              { text: 'Already up to date.', type: 'success' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git fetch ─────────────────────────────────────────────────────────
-        case 'fetch': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const remote = args.find((a) => !a.startsWith('-') && a !== sub) ?? 'origin';
-          if (!g.remotes[remote]) {
-            return { lines: [{ text: `error: '${remote}' does not appear to be a git repository.`, type: 'error' }], newState };
-          }
-          return {
-            lines: [
-              { text: `From ${g.remotes[remote]}`, type: 'output' },
-              { text: ` * branch            ${g.branch}     -> FETCH_HEAD`, type: 'output' },
-              { text: 'Fetched all remote refs.', type: 'success' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git clone ─────────────────────────────────────────────────────────
-        case 'clone': {
-          const url = args.find((a) => !a.startsWith('-') && a !== sub) ?? '';
-          if (!url) return { lines: [{ text: 'Usage: git clone <url> [directory]', type: 'error' }], newState };
-          const dirName = args.find((a) => !a.startsWith('-') && a !== sub && a !== url)
-            ?? url.split('/').pop()?.replace(/\.git$/, '') ?? 'repo';
-          newState = {
-            ...newState,
-            git: {
-              initialized: true,
-              branch: 'main',
-              branches: ['main'],
-              stagedFiles: [],
-              commits: [{ hash: makeHash(), message: 'Initial commit', author: 'remote', date: new Date().toISOString().slice(0, 10) }],
-              remotes: { origin: url },
-            },
-          };
-          return {
-            lines: [
-              { text: `Cloning into '${dirName}'...`, type: 'output' },
-              { text: 'remote: Enumerating objects: 12, done.', type: 'output' },
-              { text: 'remote: Counting objects: 100% (12/12), done.', type: 'output' },
-              { text: 'Receiving objects: 100% (12/12), done.', type: 'success' },
-              { text: `Repository cloned into '${dirName}'`, type: 'success' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git stash ─────────────────────────────────────────────────────────
-        case 'stash': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const stashSub = args[1]?.toLowerCase() ?? 'push';
-          if (stashSub === 'push' || stashSub === '') {
-            const g = newState.git!;
-            if (g.stagedFiles.length === 0) return { lines: [{ text: 'No local changes to save.', type: 'info' }], newState };
-            newState = { ...newState, git: { ...g, stagedFiles: [] } };
-            return { lines: [{ text: `Saved working directory and index state WIP on ${g.branch}: stash@{0}`, type: 'success' }], newState };
-          }
-          if (stashSub === 'pop' || stashSub === 'apply') {
-            return { lines: [{ text: 'Applied stash@{0} (simulated)', type: 'success' }], newState };
-          }
-          if (stashSub === 'list') {
-            return { lines: [{ text: 'stash@{0}: WIP on main (simulated)', type: 'output' }], newState };
-          }
-          return { lines: [{ text: 'Usage: git stash [push|pop|apply|list]', type: 'error' }], newState };
-        }
-
-        // ── git tag ───────────────────────────────────────────────────────────
-        case 'tag': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const tagName = args.find((a) => !a.startsWith('-') && a !== sub) ?? '';
-          if (!tagName) return { lines: [{ text: 'Usage: git tag <tagname>', type: 'error' }], newState };
-          const g = newState.git!;
-          if (g.commits.length === 0) return { lines: [{ text: 'fatal: No commits to tag.', type: 'error' }], newState };
-          return { lines: [{ text: `Tag '${tagName}' created at ${g.commits[0].hash}`, type: 'success' }], newState };
-        }
-
-        // ── git show ──────────────────────────────────────────────────────────
-        case 'show': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          if (g.commits.length === 0) return { lines: [{ text: 'fatal: No commits yet.', type: 'error' }], newState };
-          const c = g.commits[0];
-          return {
-            lines: [
-              { text: `commit ${c.hash}`, type: 'success' },
-              { text: `Author: ${c.author} <${c.author}@terminal-lab.local>`, type: 'output' },
-              { text: `Date:   ${c.date}`, type: 'output' },
-              { text: '', type: 'output' },
-              { text: `    ${c.message}`, type: 'output' },
-            ],
-            newState,
-          };
-        }
-
-        // ── git reset ─────────────────────────────────────────────────────────
-        case 'reset': {
-          const repoErr = requireRepo();
-          if (repoErr) return { lines: [repoErr], newState };
-          const g = newState.git!;
-          const soft = args.includes('--soft');
-          const hard = args.includes('--hard');
-          const mixed = !soft && !hard;
-          if (mixed || soft) {
-            newState = { ...newState, git: { ...g, stagedFiles: [] } };
-            return { lines: [{ text: 'Unstaged all changes (simulated).', type: 'info' }], newState };
-          }
-          if (hard) {
-            newState = { ...newState, git: { ...g, stagedFiles: [] } };
-            return { lines: [{ text: 'HEAD is now at (hard reset simulated). Working tree clean.', type: 'info' }], newState };
-          }
-          return { lines: [{ text: 'Usage: git reset [--soft|--hard] [<commit>]', type: 'error' }], newState };
-        }
-
-        // ── git config ────────────────────────────────────────────────────────
-        case 'config': {
-          const globalFlag = args.includes('--global');
-          const key = args.find((a) => !a.startsWith('-') && a !== sub) ?? '';
-          const value = args[args.indexOf(key) + 1] ?? '';
-          if (key === 'user.name' || key === 'user.email') {
-            return { lines: [{ text: `${key} = ${value || newState.user} (configuré)`, type: 'success' }], newState };
-          }
-          if (key === '--list' || args.includes('--list')) {
-            return {
-              lines: [
-                { text: `user.name=${newState.user}`, type: 'output' },
-                { text: `user.email=${newState.user}@terminal-lab.local`, type: 'output' },
-                { text: 'core.editor=nano', type: 'output' },
-                { text: `init.defaultBranch=main`, type: 'output' },
-                ...(globalFlag ? [{ text: 'credential.helper=store', type: 'output' as const }] : []),
-              ],
-              newState,
-            };
-          }
-          return { lines: [{ text: `git config ${key || '--list'}`, type: 'info' }], newState };
-        }
-
-        // ── git --version / --help ────────────────────────────────────────────
-        case '--version':
-          return { lines: [{ text: 'git version 2.45.0 (simulated)', type: 'output' }], newState };
-
-        case '--help':
-        case 'help':
-          return {
-            lines: [
-              { text: 'usage: git <command> [<args>]', type: 'output' },
-              { text: '', type: 'output' },
-              { text: 'Commandes essentielles :', type: 'info' },
-              { text: '  init      Initialiser un nouveau dépôt', type: 'output' },
-              { text: '  clone     Cloner un dépôt distant', type: 'output' },
-              { text: '  add       Ajouter des fichiers à l\'index (staging)', type: 'output' },
-              { text: '  commit    Enregistrer les modifications indexées', type: 'output' },
-              { text: '  status    Afficher l\'état du répertoire de travail', type: 'output' },
-              { text: '  log       Afficher l\'historique des commits', type: 'output' },
-              { text: '  diff      Comparer les modifications', type: 'output' },
-              { text: '  branch    Lister, créer ou supprimer des branches', type: 'output' },
-              { text: '  checkout  Changer de branche ou restaurer des fichiers', type: 'output' },
-              { text: '  switch    Changer de branche (commande moderne)', type: 'output' },
-              { text: '  merge     Fusionner une branche dans la branche active', type: 'output' },
-              { text: '  remote    Gérer les dépôts distants', type: 'output' },
-              { text: '  push      Envoyer les commits vers le dépôt distant', type: 'output' },
-              { text: '  pull      Récupérer et intégrer les commits distants', type: 'output' },
-              { text: '  fetch     Récupérer sans intégrer (pull sans merge)', type: 'output' },
-              { text: '  stash     Mettre de côté des modifications temporairement', type: 'output' },
-              { text: '  tag       Créer un tag sur un commit', type: 'output' },
-              { text: '  show      Afficher un commit en détail', type: 'output' },
-              { text: '  reset     Annuler des modifications', type: 'output' },
-              { text: '  config    Configurer git (user.name, user.email…)', type: 'output' },
-            ],
-            newState,
-          };
-
-        default:
-          return {
-            lines: [{
-              text: `git: '${sub}' is not a git command. See 'git help'.`,
-              type: 'error',
-            }],
-            newState,
-          };
-      }
-    }
-
-    default:
+    // ── Windows/macOS aliases & platform commands → commands/windows.ts ───────
+    default: {
+      const winResult = handleWindows(cmd, args, newState, env, winDeps);
+      if (winResult !== null) return winResult;
       return {
         lines: [{ text: `${cmd}: commande introuvable. Tapez 'help' pour la liste des commandes.`, type: 'error' }],
         newState,
       };
+    }
   }
 }
